Convert page layout to use CSS flexbox
This is basically just to keep the footer hugging the bottom of the
viewport regardless of what's on the page.

The .page div is a flex container, and all its child elements flex
items. The child .main div will grow to fill vertical space, so that
footer is always pushed to the bottom (and won't ever be shrunk by this
resizing).

Blog pages have also been changed to match this three-part structure,
with .header, .main and footer elements inheriting these flexbox
properties from main stylesheet.